@@ -1,3 +1,6 @@
+// imports
+import { v4 as uuidv4 } from 'uuid';
+
 //! firebase
 const { initializeApp } = require('firebase/app');
 const { getStorage, ref, uploadBytes, getDownloadURL } = require('firebase/storage');
@@ -8,46 +11,28 @@ const storage = getStorage();
 
 
 
-//**********   getNoticePDFs      *************  */
+//*** Method to upload a file to Firebase Storage and return the download URL  *****  */
 
-const SingleNoticePDFUrl = async (notice: any) => {
-    const storage = getStorage();
+export const uploadFileToFirebaseAndGetDownloadUrl = async (
+    uuid: string,
+    accountId: string,
+    file: Express.Multer.File,
+): Promise<string> => {
     try {
-        const pdfRef = ref(storage, `notice/pdf/${notice.pdf}`);
-        const url = await getDownloadURL(pdfRef);
-        return url;
+
+        const originalFilename = file.originalname;
+        const mimetype = file.mimetype;
+
+        const filename = `${accountId}-${uuid}-${originalFilename}`;
+        const metadata = { contentType: mimetype };
+        const storage = getStorage(); // Get a reference to the Firebase Storage bucket
+        const pdfRef = ref(storage, `notice/academyId-${accountId}/pdf/${filename}`); // Create a reference to the bucket
+        await uploadBytes(pdfRef, file.buffer, metadata);
+        const pdfUrl: string = await getDownloadURL(pdfRef);
+
+        return pdfUrl;
     } catch (error) {
-        return "";
+        console.error('Error uploading file to Firebase:', error);
+        throw error;
     }
-
-
-};
-
-export const getNoticePDFs = async (notices: any) => {
-    for (let i = 0; i < notices.length; i++) {
-        try {
-            const notice = notices[i];
-            const pdfUrls = await SingleNoticePDFUrl(notice);
-            notice.pdf = pdfUrls;
-        } catch (err) {
-            console.error(err);
-        }
-    }
-    return notices;
-};
-
-
-
-export const getNoticePDFUrls = async (notice: any) => {
-    const urls = [];
-    const storage = getStorage();
-
-    // Loop through all the PDFs in the notice and get their download URLs
-    for (let i = 0; i < notice.pdf.length; i++) {
-        const pdfRef = ref(storage, `notice/pdf/${notice.pdf[i].url}`);
-        const url = await getDownloadURL(pdfRef);
-        urls.push({ url, _id: notice.pdf[i]._id });
-    }
-
-    return urls;
 };
